feature/refactor-easy-utils
- Refactor EasyUtils
